Move percentile to array
